+++ modules/home/scripts/default.nix
@@ -17,5 +17,6 @@
     (import ./rofi-launcher.nix {inherit pkgs;})
     (import ./screenshootin.nix {inherit pkgs;})
     (import ./hm-find.nix {inherit pkgs;})
+    (import ./gemini.nix {inherit pkgs;})
   ];
 }