+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## ZaneyOS v2.3.3 -- Post GA Release Notes
 
-** Updated: June 25th, 2025 **
+** Updated: June 27th, 2025 **
 
+- Added `gemini.nix` Creates an `npm` script to install `gemini-cli`
 - Added `evil-helix` with language support as option
 - Added `vscode.nix` with plugins for NIX, BASH, JSON, and VIM keybinds
 - Thanks to `delciak` for providing the NIX code for `vscode.nix`